0.23.6 Huge code cleanup and some restructuring

+++ src/models/selection_model.py
@@ -7,11 +7,12 @@ from sprites import CardSprite
 
 
 class SelectionModel(Model):
-    def __init__(self, card_image_sheet=None):
+    def __init__(self, card_image_sheet=None, selection_manager=None):
         if card_image_sheet is None:
             raise RuntimeError("SelectionModel card_image_sheet cannot be None")
+        if selection_manager is None:
+            raise RuntimeError("SelectionModel selection_manager cannot be None")
 
-        self._card_image_sheet = card_image_sheet
         self._pile = Pile(rect=(0, 0, 80, 120))
 
         def sync_pile_rect_with_mouse():
@@ -21,49 +22,29 @@ class SelectionModel(Model):
             pile_rect.y = mouse_y - 20
             return
 
-        self._mouse_sync_element = Element(on_update=sync_pile_rect_with_mouse)
-        self._pile_element = PileElement(pile=self._pile, direction=PileElement.PileElementDirection.DOWN, spacing=20)
-        self._last_selected_collection = None
-        self._card_sprite_to_show = None
-
-        def undo_selection():
-            if self.is_empty():
-                return
-            selected_cards = self.remove_all()
-            last_selected_collection = self.get_last_selected_collection()
-            for card in selected_cards:
-                last_selected_collection.insert(card)
+        def sync_pile_sprites_with_collection():
+            card_collection = selection_manager.get_card_collection()
+            if len(self._pile) != len(card_collection):
+                self._pile.remove_all()
+                for card in card_collection.get_cards():
+                    card_sprite = CardSprite(card=card, rect=(0, 0, 0, 0),
+                                             card_image_sheet=card_image_sheet)
+                    self._pile.add(card_sprite)
             return
 
+        def undo_selection():
+            selection_manager.undo_selection()
+            self._pile.remove_all()
+            return
+
+        self._sprite_sync_element = Element(on_update=sync_pile_sprites_with_collection)
+        self._mouse_sync_element = Element(on_update=sync_pile_rect_with_mouse)
+        self._pile_element = PileElement(pile=self._pile, direction=PileElement.PileElementDirection.DOWN, spacing=20)
         self._mouse_handler = MouseHandler(on_left_mouse_release=undo_selection)
         return
 
-    def is_empty(self):
-        return self._pile.is_empty()
-
-    def __len__(self):
-        return len(self._pile.get_entries())
-
-    def get_last_selected_collection(self):
-        return self._last_selected_collection
-
-    def set_last_selected_collection(self, card_collection):
-        self._last_selected_collection = card_collection
-        return
-
-    def add_card(self, card, shown=True):
-        card_sprite = CardSprite(card=card, rect=(0, 0, 0, 0),
-                                 card_image_sheet=self._card_image_sheet, shown=shown)
-        self._pile.add(card_sprite)
-        return
-
-    def remove_all(self):
-        card_sprites = self._pile.remove_all()
-        cards = [card_sprite.get_card() for card_sprite in card_sprites]
-        return cards
-
     def get_elements(self):
-        return tuple([self._mouse_sync_element, self._pile_element])
+        return tuple([self._sprite_sync_element, self._mouse_sync_element, self._pile_element])
 
     def get_sprites(self):
         return self._pile.get_entries()
